dev-requirements.txt
pytest
pytest-cov
setuptools
cython 
# pysam
requests 
coverage 
pyfasta 
pyvcf3
numpy 
biopython
fsspec
